backend/pyproject.toml
[project]
name = "mirmer-ai-backend"
version = "0.1.0"
description = "Multi-LLM consultation system backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "sendgrid>=6.11.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "hypothesis>=6.92.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
